refactor : domain패키지 분리
