+++ src/main/java/cart/persistence/H2MemberDao.java
@@ -1,7 +1,7 @@
 package cart.persistence;
 
-import cart.service.member.Member;
 import cart.service.member.MemberDao;
+import cart.service.member.domain.Member;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
