+++ src/main/java/cart/service/member/MemberDao.java
@@ -1,5 +1,7 @@
 package cart.service.member;
 
+import cart.service.member.domain.Member;
+
 import java.util.List;
 import java.util.Optional;
