+++ src/main/java/cart/service/product/domain/ProductName.java
@@ -1,4 +1,4 @@
-package cart.service.product;
+package cart.service.product.domain;
 
 public class ProductName {
     private static final int MAX_NAME_LENGTH = 50;
