+++ src/test/java/cart/service/product/ProductPriceTest.java
@@ -1,5 +1,6 @@
 package cart.service.product;
 
+import cart.service.product.domain.ProductPrice;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
